ビューの継承

+++ resources/views/posts/comment.blade.php
@@ -1,25 +1,9 @@
-<!DOCTYPE html>
-<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-
-        <title>main</title>
-
-        <!-- Fonts -->
-        <link href="https://fonts.bunny.net/css2?family=Nunito:wght@400;600;700&display=swap" rel="stylesheet">
-        
-         <!-- CSSの読み込み -->
-        <link rel="stylesheet" href="../css/style.css">
-        
-    </head>
-    <body>
-    <div id="page">
-    <header>
+<x-app-layout>
+    <x-slot name="header">
         <h1>口コミ一覧</h1>
-    </header>
+    </x-slot>
     
-    <main>
+    <x-slot name="slot">
         <div class='comments'>
             @foreach ($comments as $comment)
                 <div class='comment'>
@@ -28,14 +12,17 @@
                     <p class='body'>{{ $comment->body }}</p>
                     <p class='time'>{{ $comment->created_at }}</p>
                 </div>
+                @if(Auth::id() == $comment->user_id){
+                <form action="/posts/{{ $comment->id }}" id="form_{{ $comment->id }}" method="post">
+                    @csrf
+                    @method('DELETE')
+                    <button type="button" onclick="deletePost({{ $comment->id }})">delete</button> 
+                </form>
+                <div class="edit"><a href="/posts/{{ $comment->id }}/edit">edit</a></div>
+                }
+                @endif
             @endforeach
         </div>
-        <form action="/posts/{{ $comment->id }}" id="form_{{ $comment->id }}" method="post">
-                @csrf
-                @method('DELETE')
-                <button type="button" onclick="deletePost({{ $comment->id }})">delete</button> 
-         </form>
-        <div class="edit"><a href="/posts/{{ $comment->id }}/edit">edit</a></div>
         <script>
             function deletePost(id) {
                 'use strict'
@@ -44,33 +31,35 @@
                 }
             }
         </script>
-    </main>
+    </x-slot>
     
-    <div id="sub1">
-        <h1>アニメを探す</h1>
-        <a href='/posts/anime'>roll</a>
-        <ul id="sample1">
-            <li>ランキング</li>
-            <li>年代別アニメ</li>
-            <li>ジャンル別アニメ</li>
-        </ul>
-        <h1>アニメ評価</h1>
-        <h1>アニメ一覧</h1>
-        <a href='/posts/anime'>roll</a>
-        <h1>オススメアニメ投稿</h1>
-        <a href='/posts/create'>create</a>
-        <h1>口コミ一覧</h1>
-        <a href='/posts/comment'>roll2</a>
-        <h1>口コミ投稿</h1>
-        <a href='/posts/comment_create'>create_c</a>
-        
-    </div>
+    <x-slot name="sub1">
+        <div id="sub1">
+            <h1>アニメを探す</h1>
+            <a href='/posts/anime'>roll</a>
+            <ul id="sample1">
+                <li>ランキング</li>
+                <li>年代別アニメ</li>
+                <li>ジャンル別アニメ</li>
+            </ul>
+            <h1>アニメ評価</h1>
+            <h1>アニメ一覧</h1>
+            <a href='/posts/anime'>roll</a>
+            <h1>オススメアニメ投稿</h1>
+            <a href='/posts/create'>create</a>
+            <h1>口コミ一覧</h1>
+            <a href='/posts/comment'>roll2</a>
+            <h1>口コミ投稿</h1>
+            <a href='/posts/comment_create'>create_c</a>
+        </div>
+    </x-slot>
     
-    <div id="sub2">
-        <h1>今期のアニメ一覧</h1>
-        <h1>新規の口コミ</h1>
-    </div>
+    <x-slot name="sub2">
+        <div id="sub2">
+            <h1>今期のアニメ一覧</h1>
+            <h1>新規の口コミ</h1>
+        </div>
+     </x-slot>
+ </x-app-layout> 
     
-    </div>
-    </body>
-</html>
+
